Controller changes, start of data handling

+++ src/main/java/tk/chaber/triangles_properties/Geometry/Angle.java
@@ -4,7 +4,7 @@ public class Angle {
     private double radians;
     private double degrees;
 
-    public Angle(int degrees){
+    public Angle(double degrees){
         this.degrees = degrees;
         this.radians = degrees * Math.PI/180;
     }
